+++ include/votca/csg/topology.h
@@ -77,7 +77,8 @@ class Topology {
    * The function creates a new bead and adds it to the list of beads.
    */
   virtual Bead *CreateBead(tools::byte_t symmetry, std::string name,
-                           std::string type, int resnr, double m, double q);
+                           std::string type, long int resnr, double m,
+                           double q);
 
   /**
    * \brief Creates a new molecule
@@ -99,7 +100,7 @@ class Topology {
    * @return created residue
    */
   virtual Residue *CreateResidue(std::string name);
-  virtual Residue *CreateResidue(std::string name, int id);
+  virtual Residue *CreateResidue(std::string name, long int id);
 
   /**
    * \brief Create molecules based on the residue.
@@ -124,33 +125,33 @@ class Topology {
    * \param[in] nbeads number of beads per molecule
    * \param[in] nmolecules number of molecules
    */
-  void CreateMoleculesByRange(std::string name, int first, int nbeads,
-                              int nmolecules);
+  void CreateMoleculesByRange(std::string name, long int first, long int nbeads,
+                              long int nmolecules);
 
   /**
    * \brief number of molecules in the system
    * @return number of molecule in topology
    */
-  int MoleculeCount() const { return _molecules.size(); }
+  long int MoleculeCount() const { return _molecules.size(); }
 
   /**
    * number of beads in the system
    * @return number of beads in the system
    */
-  int BeadCount() const { return _beads.size(); }
+  long int BeadCount() const { return _beads.size(); }
 
   /**
    * number of residues in the system
    * \return number of residues
    */
-  int ResidueCount() const { return _residues.size(); }
+  long int ResidueCount() const { return _residues.size(); }
 
   /**
    * get molecule by index
    * @param index molecule number
    * @return pointer to molecule
    */
-  Molecule *MoleculeByIndex(int index);
+  Molecule *MoleculeByIndex(long int index);
 
   /**
    * access containter with all beads
@@ -202,7 +203,7 @@ class Topology {
    * @param[in] string name of the type
    * @return int the id of the type
    **/
-  int getBeadTypeId(std::string type) const;
+  long int getBeadTypeId(std::string type) const;
 
   /**
    * \brief Returns a pointer to the bead with index i
@@ -210,9 +211,9 @@ class Topology {
    * @param[in] int i is the id of the bead
    * @return Bead * is a pointer to the bead
    **/
-  Bead *getBead(const int i) const { return _beads[i]; }
-  Residue *getResidue(const int i) const { return _residues[i]; }
-  Molecule *getMolecule(const int i) const { return _molecules[i]; }
+  Bead *getBead(const long int i) const { return _beads[i]; }
+  Residue *getResidue(const long int i) const { return _residues[i]; }
+  Molecule *getMolecule(const long int i) const { return _molecules[i]; }
 
   /**
    * delete all molecule information
@@ -305,13 +306,13 @@ class Topology {
    * set the step number of current frame
    * \param s step number
    */
-  void setStep(int s) { _step = s; };
+  void setStep(long int s) { _step = s; };
 
   /**
    * get the step number of current frame
    * \return step number
    */
-  int getStep() const { return _step; };
+  long int getStep() const { return _step; };
 
   /**
    * Sets the particle group. (For the H5MD file format)
@@ -336,7 +337,7 @@ class Topology {
    * calculates the smallest distance between two beads with correct treatment
    * of pbc
    */
-  Eigen::Vector3d getDist(int bead1, int bead2) const;
+  Eigen::Vector3d getDist(long int bead1, long int bead2) const;
 
   /**
    * \brief calculate shortest vector connecting two points
@@ -393,7 +394,7 @@ class Topology {
       const Eigen::Matrix3d &box) const;
 
   /// bead types in the topology
-  std::unordered_map<std::string, int> beadtypes_;
+  std::unordered_map<std::string, long int> beadtypes_;
 
   /// beads in the topology
   BeadContainer _beads;
@@ -409,12 +410,12 @@ class Topology {
 
   ExclusionList _exclusions;
 
-  std::map<std::string, int> _interaction_groups;
+  std::map<std::string, long int> _interaction_groups;
 
   std::map<std::string, std::list<Interaction *> > _interactions_by_group;
 
   double _time = 0.0;
-  int _step = 0;
+  long int _step = 0;
   bool _has_vel = false;
   bool _has_force = false;
 
@@ -423,7 +424,7 @@ class Topology {
 };
 
 inline Bead *Topology::CreateBead(tools::byte_t symmetry, std::string name,
-                                  std::string type, int resnr, double m,
+                                  std::string type, long int resnr, double m,
                                   double q) {
 
   Bead *b = new Bead(this, _beads.size(), type, symmetry, name, resnr, m, q);
@@ -437,7 +438,7 @@ inline Molecule *Topology::CreateMolecule(std::string name) {
   return mol;
 }
 
-inline Residue *Topology::CreateResidue(std::string name, int id) {
+inline Residue *Topology::CreateResidue(std::string name, long int id) {
   Residue *res = new Residue(this, id, name);
   _residues.push_back(res);
   return res;
@@ -449,7 +450,7 @@ inline Residue *Topology::CreateResidue(std::string name) {
   return res;
 }
 
-inline Molecule *Topology::MoleculeByIndex(int index) {
+inline Molecule *Topology::MoleculeByIndex(long int index) {
   return _molecules[index];
 }
